Now using Goertzel to estimate deviation
The Goertzel single bin DFT argorithm is now used to estimate the FM
deviation for a tone on a single frequency which gives a much better
estimate since much noise will be suppressed.
There also is a wideband mode (the old mode) which can be used when
measuring wideband signals like speech.